chris_lux2/requirements.txt
Django
mysqlclient>=2.2.0
Pillow>=11.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.0
django-storages 
boto3